[stream] pr 会出错 #9031

+++ src/backend/ci/core/common/common-webhook/api-common-webhook/src/main/kotlin/com/tencent/devops/common/webhook/pojo/code/github/GithubPullRequestEvent.kt
@@ -98,7 +98,7 @@ data class GithubPullRequest(
     @JsonProperty("html_url")
     val htmlUrl: String, // https://github.com/yongyiduan/webhook-test/pull/1
     @JsonProperty("id")
-    val id: Int, // 973279061
+    val id: Long, // 973279061
     @JsonProperty("issue_url")
     val issueUrl: String, // https://api.github.com/repos/yongyiduan/webhook-test/issues/1
     @JsonProperty("labels")
